requirements.txt
streamlit=1.39.0
google-generativeai
st-pages